Bunny Rework
Added coordinates to the bunny, that are changed if the coordinates are already in use. Made modifications to the mergeSort() function in order to make it accept a callback function as a parameter (to change the comparator and sort bunnies in order of their coordinates instead of IDs). Reworked the printing in order to print in accordance with the grid locations. Also modified the printing in order to make everything print at once, reducing the blinking on the console.